+++ Distances.h
@@ -2,27 +2,26 @@
 #ifndef AP1_DISTANCES_H
 #define AP1_DISTANCES_H
 #include <vector>
-
 using namespace std;
 
 class Distances {
     //members
-    vector<long double> v1, v2;
+    vector<double> v1, v2;
 public:
     //constructor
-    Distances(vector<long double> &v1, const vector<long double> &v2);
+    Distances(vector<double> &v1, const vector<double> &v2);
     //distance functions
-    long double euclidianDistance();
+    double euclidianDistance();
 
-    long double manhattenDistance();
+    double manhattenDistance();
 
-    long double chebyshevDistance();
+    double chebyshevDistance();
 
-    long double canberraDistance();
+    double canberraDistance();
 
-    long double minkowskiDistance();
+    double minkowskiDistance();
 
-    long double minkowskiDistanceGeneric(int exp);
+    double minkowskiDistanceGeneric(int exp);
 };
 
 #endif //AP1_DISTANCES_H
